% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tombRaiderWrapper.R
\name{run_tombRaider_custom}
\alias{run_tombRaider_custom}
\title{Run tombRaider with custom parameters}
\usage{
run_tombRaider_custom(
  method,
  frequency_input,
  sequence_input,
  blast_input,
  frequency_output,
  sequence_output,
  blast_output,
  condensed_log,
  detailed_log,
  count,
  sort
)
}
\arguments{
\item{method}{Method to use}

\item{frequency_input}{Path to frequency input file}

\item{sequence_input}{Path to sequence input file}

\item{blast_input}{Path to BLAST input file}

\item{frequency_output}{Path to frequency output file}

\item{sequence_output}{Path to sequence output file}

\item{blast_output}{Path to BLAST output file}

\item{condensed_log}{Path to condensed log file}

\item{detailed_log}{Path to detailed log file}

\item{count}{Count parameter}

\item{sort}{Sort parameter}
}
\description{
Run tombRaider with custom parameters
}
