% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tombRaiderWrapper.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Set up the Python environment and install required packages}
\usage{
setup_python_env()
}
\description{
Set up the Python environment and install required packages
}
