% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tombRaiderWrapper.R
\name{run_tombRaider}
\alias{run_tombRaider}
\title{Run tombRaider with specified arguments}
\usage{
run_tombRaider(args)
}
\arguments{
\item{args}{Command line arguments for tombRaider}
}
\description{
Run tombRaider with specified arguments
}
